Move check funtions to method.utils.py
